refactor(TNLT-6524): hide close button on android for tooltip (#320)
* refactor: hide close button on android for tooltip

* fix: added test to hide close button on android only for tooltip

* fix: snaps

* fix: all tooltip tests and split out into relevant directory
